Pracenje blizine aktivnosti izgleda da radi Konacno smo dosli do kraja Sad ide testiranje i popravke za pred odbranu

+++ CampingApp/app/src/main/java/elfak/mosis/campingapp/fragments/FragmentTripTeammatesMap.kt
@@ -150,8 +150,8 @@ class FragmentTripTeammatesMap : Fragment()
                 }
 
 
-                Location.distanceBetween(l?.latitude ?: 0 as Double, l?.longitude ?: 0 as Double, aktivnost.latitude, aktivnost.longitude, tmp)
-                if (sharedViewModel.distance.value!! > 0 && tmp[0] / 1000 > sharedViewModel.distance.value!!)
+                Location.distanceBetween(l?.latitude ?: 0.toDouble(), l?.longitude ?: 0.toDouble(), aktivnost.latitude, aktivnost.longitude, tmp)
+                if (sharedViewModel.distance.value!! > 0.toDouble() && tmp[0].toDouble() / 1000 > sharedViewModel.distance.value!!)
                     continue
 
                 var marker = Marker(mapa)
